pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "life-world-model"
version = "0.1.0"
description = "A local-first experiment that turns desktop activity into a Tolkien-esque day narrative"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["world-models", "narrative", "lifelogging", "cli"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Utilities"
]
dependencies = []

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]
mlx = [
  "mlx-lm>=0.21.0"
]
gemini = [
  "google-genai>=1.0.0"
]
web = [
  "flask>=3.0.0"
]

[project.scripts]
lwm = "life_world_model.cli:main"
lwm-collect = "life_world_model.cli:collect_entrypoint"
lwm-generate = "life_world_model.cli:generate_entrypoint"
lwm-run = "life_world_model.cli:run_entrypoint"

[tool.hatch.build.targets.wheel]
packages = ["src/life_world_model"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
